organize mcp env stuff
