+++ verifiers/__init__.py
@@ -12,7 +12,7 @@ from .envs.multiturn_env import MultiTurnEnv
 from .envs.singleturn_env import SingleTurnEnv
 from .envs.stateful_tool_env import StatefulToolEnv
 from .envs.tool_env import ToolEnv
-from .envs.mcp.mcp_env import MCPEnv
+from .envs.mcp_env import MCPEnv
 from .parsers.maybe_think_parser import MaybeThinkParser
 from .parsers.parser import Parser
 from .parsers.think_parser import ThinkParser
